back_end/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cloudinary==1.41.0
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.34.0
whitenoise==6.8.2
